PhenotypeSpace: Quantifying phenotypic spaces

Facilitates quantifying phenotypic trait spaces and comparing spaces across groups.

Details

The package offers functions for:

Estimate absolute values of phenotypic trait spaces

Compare phenotypic trait spaces across groups

Modify output formats to facilitate statistical analysis

All functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency.

License: GPL (>= 2)

See also

Author

Marcelo Araya-Salas

Maintainer: Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)